+++ grovedb/src/batch/average_case_costs.rs
@@ -30,6 +30,14 @@ pub(super) struct AverageCaseTreeCacheKnownPaths {
     paths: HashMap<KeyInfoPath, EstimatedLayerInformation>,
 }
 
+impl AverageCaseTreeCacheKnownPaths {
+    pub(super) fn new_with_estimated_layer_information(
+        paths: HashMap<KeyInfoPath, EstimatedLayerInformation>,
+    ) -> Self {
+        AverageCaseTreeCacheKnownPaths { paths }
+    }
+}
+
 impl fmt::Debug for AverageCaseTreeCacheKnownPaths {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_struct("TreeCacheKnownPaths").finish()
@@ -114,5 +122,347 @@ impl<G, SR> TreeCache<G, SR> for AverageCaseTreeCacheKnownPaths {
         } else {
         }
         Ok(()).wrap_with_cost(cost)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::collections::HashMap;
+
+    use costs::{
+        storage_cost::{removal::StorageRemovedBytes::NoStorageRemoval, StorageCost},
+        OperationCost,
+    };
+    use merk::estimated_costs::average_case_costs::{
+        EstimatedLayerInformation,
+        EstimatedLayerInformation::EstimatedLevel,
+        EstimatedLayerSizes::{AllItems, AllSubtrees},
+    };
+
+    use crate::{
+        batch::{
+            estimated_costs::EstimatedCostsType::AverageCaseCostsType, key_info::KeyInfo,
+            GroveDbOp, KeyInfoPath,
+        },
+        tests::make_empty_grovedb,
+        Element, GroveDb,
+    };
+
+    #[test]
+    fn test_batch_root_one_tree_insert_op_average_case_costs() {
+        let db = make_empty_grovedb();
+        let tx = db.start_transaction();
+
+        let ops = vec![GroveDbOp::insert_op(
+            vec![],
+            b"key1".to_vec(),
+            Element::empty_tree(),
+        )];
+        let mut paths = HashMap::new();
+        paths.insert(KeyInfoPath(vec![]), EstimatedLevel(0, AllSubtrees(4, None)));
+        paths.insert(
+            KeyInfoPath(vec![KeyInfo::KnownKey(b"key1".to_vec())]),
+            EstimatedLevel(0, AllSubtrees(4, None)),
+        );
+        let average_case_cost = GroveDb::estimated_case_operations_for_batch(
+            AverageCaseCostsType(paths),
+            ops.clone(),
+            None,
+            |_cost, _old_flags, _new_flags| Ok(false),
+            |_flags, _removed_key_bytes, _removed_value_bytes| {
+                Ok((NoStorageRemoval, NoStorageRemoval))
+            },
+        )
+        .cost_as_result()
+        .expect("expected to get average case costs");
+
+        let cost = db.apply_batch(ops, None, Some(&tx)).cost;
+        assert!(
+            average_case_cost.worse_or_eq_than(&cost),
+            "not worse {:?} \n than {:?}",
+            average_case_cost,
+            cost
+        );
+        // because we know the object we are inserting we can know the average
+        // case cost if it doesn't already exist
+        assert_eq!(
+            cost.storage_cost.added_bytes,
+            average_case_cost.storage_cost.added_bytes
+        );
+
+        assert_eq!(
+            average_case_cost,
+            OperationCost {
+                seek_count: 6, // todo: why is this 6
+                storage_cost: StorageCost {
+                    added_bytes: 113,
+                    replaced_bytes: 18432, // todo: verify
+                    removed_bytes: NoStorageRemoval,
+                },
+                storage_loaded_bytes: 23040,
+                hash_node_calls: 18, // todo: verify why
+            }
+        );
+    }
+
+    #[test]
+    fn test_batch_root_one_tree_with_flags_insert_op_average_case_costs() {
+        let db = make_empty_grovedb();
+        let tx = db.start_transaction();
+
+        let ops = vec![GroveDbOp::insert_op(
+            vec![],
+            b"key1".to_vec(),
+            Element::empty_tree_with_flags(Some(b"cat".to_vec())),
+        )];
+        let mut paths = HashMap::new();
+        paths.insert(
+            KeyInfoPath(vec![]),
+            EstimatedLevel(0, AllSubtrees(4, Some(3))),
+        );
+        let average_case_cost = GroveDb::estimated_case_operations_for_batch(
+            AverageCaseCostsType(paths),
+            ops.clone(),
+            None,
+            |_cost, _old_flags, _new_flags| Ok(false),
+            |_flags, _removed_key_bytes, _removed_value_bytes| {
+                Ok((NoStorageRemoval, NoStorageRemoval))
+            },
+        )
+        .cost_as_result()
+        .expect("expected to get average case costs");
+
+        let cost = db.apply_batch(ops, None, Some(&tx)).cost;
+        assert!(
+            average_case_cost.worse_or_eq_than(&cost),
+            "not worse {:?} \n than {:?}",
+            average_case_cost,
+            cost
+        );
+        // because we know the object we are inserting we can know the average
+        // case cost if it doesn't already exist
+        assert_eq!(
+            cost.storage_cost.added_bytes,
+            average_case_cost.storage_cost.added_bytes
+        );
+
+        assert_eq!(
+            average_case_cost,
+            OperationCost {
+                seek_count: 6, // todo: why is this 6
+                storage_cost: StorageCost {
+                    added_bytes: 117,
+                    replaced_bytes: 18432, // todo: verify
+                    removed_bytes: NoStorageRemoval,
+                },
+                storage_loaded_bytes: 23040,
+                hash_node_calls: 18, // todo: verify why
+            }
+        );
+    }
+
+    #[test]
+    fn test_batch_root_one_item_insert_op_average_case_costs() {
+        let db = make_empty_grovedb();
+        let tx = db.start_transaction();
+
+        let ops = vec![GroveDbOp::insert_op(
+            vec![],
+            b"key1".to_vec(),
+            Element::new_item(b"cat".to_vec()),
+        )];
+        let mut paths = HashMap::new();
+        paths.insert(KeyInfoPath(vec![]), EstimatedLevel(0, AllItems(4, 3, None)));
+        let average_case_cost = GroveDb::estimated_case_operations_for_batch(
+            AverageCaseCostsType(paths),
+            ops.clone(),
+            None,
+            |_cost, _old_flags, _new_flags| Ok(false),
+            |_flags, _removed_key_bytes, _removed_value_bytes| {
+                Ok((NoStorageRemoval, NoStorageRemoval))
+            },
+        )
+        .cost_as_result()
+        .expect("expected to get average case costs");
+
+        let cost = db.apply_batch(ops, None, Some(&tx)).cost;
+        assert!(
+            average_case_cost.worse_or_eq_than(&cost),
+            "not worse {:?} \n than {:?}",
+            average_case_cost,
+            cost
+        );
+        // because we know the object we are inserting we can know the average
+        // case cost if it doesn't already exist
+        assert_eq!(
+            cost.storage_cost.added_bytes,
+            average_case_cost.storage_cost.added_bytes
+        );
+
+        assert_eq!(
+            average_case_cost,
+            OperationCost {
+                seek_count: 4, // todo: why is this 6
+                storage_cost: StorageCost {
+                    added_bytes: 147,
+                    replaced_bytes: 18432, // log(max_elements) * 32 = 640 // todo: verify
+                    removed_bytes: NoStorageRemoval,
+                },
+                storage_loaded_bytes: 23040,
+                hash_node_calls: 18, // todo: verify why
+            }
+        );
+    }
+
+    #[test]
+    fn test_batch_root_one_tree_insert_op_under_element_average_case_costs() {
+        let db = make_empty_grovedb();
+        let tx = db.start_transaction();
+
+        db.insert(vec![], b"0", Element::empty_tree(), None, Some(&tx))
+            .unwrap()
+            .expect("successful root tree leaf insert");
+
+        let ops = vec![GroveDbOp::insert_op(
+            vec![],
+            b"key1".to_vec(),
+            Element::empty_tree(),
+        )];
+        let mut paths = HashMap::new();
+        paths.insert(KeyInfoPath(vec![]), EstimatedLevel(1, AllSubtrees(2, None)));
+        let average_case_cost = GroveDb::estimated_case_operations_for_batch(
+            AverageCaseCostsType(paths),
+            ops.clone(),
+            None,
+            |_cost, _old_flags, _new_flags| Ok(false),
+            |_flags, _removed_key_bytes, _removed_value_bytes| {
+                Ok((NoStorageRemoval, NoStorageRemoval))
+            },
+        )
+        .cost_as_result()
+        .expect("expected to get average case costs");
+
+        let cost = db.apply_batch(ops, None, Some(&tx)).cost;
+        assert!(
+            average_case_cost.worse_or_eq_than(&cost),
+            "not worse {:?} \n than {:?}",
+            average_case_cost,
+            cost
+        );
+        // because we know the object we are inserting we can know the average
+        // case cost if it doesn't already exist
+        assert_eq!(
+            cost.storage_cost.added_bytes,
+            average_case_cost.storage_cost.added_bytes
+        );
+
+        assert_eq!(
+            average_case_cost,
+            OperationCost {
+                seek_count: 6, // todo: why is this 6
+                storage_cost: StorageCost {
+                    added_bytes: 113,
+                    replaced_bytes: 18432, // todo: verify
+                    removed_bytes: NoStorageRemoval,
+                },
+                storage_loaded_bytes: 23040,
+                hash_node_calls: 18, // todo: verify why
+            }
+        );
+    }
+
+    #[test]
+    fn test_batch_root_one_tree_insert_op_in_sub_tree_average_case_costs() {
+        let db = make_empty_grovedb();
+        let tx = db.start_transaction();
+
+        db.insert(vec![], b"0", Element::empty_tree(), None, Some(&tx))
+            .unwrap()
+            .expect("successful root tree leaf insert");
+
+        let ops = vec![GroveDbOp::insert_op(
+            vec![b"0".to_vec()],
+            b"key1".to_vec(),
+            Element::empty_tree(),
+        )];
+        let mut paths = HashMap::new();
+        paths.insert(KeyInfoPath(vec![]), EstimatedLevel(0, AllSubtrees(1, None)));
+        paths.insert(
+            KeyInfoPath(vec![KeyInfo::KnownKey(b"0".to_vec())]),
+            EstimatedLevel(0, AllSubtrees(4, None)),
+        );
+        let average_case_cost = GroveDb::estimated_case_operations_for_batch(
+            AverageCaseCostsType(paths),
+            ops.clone(),
+            None,
+            |_cost, _old_flags, _new_flags| Ok(false),
+            |_flags, _removed_key_bytes, _removed_value_bytes| {
+                Ok((NoStorageRemoval, NoStorageRemoval))
+            },
+        )
+        .cost_as_result()
+        .expect("expected to get average case costs");
+
+        let cost = db.apply_batch(ops, None, Some(&tx)).cost;
+        assert!(
+            average_case_cost.worse_or_eq_than(&cost),
+            "not worse {:?} \n than {:?}",
+            average_case_cost,
+            cost
+        );
+        // /// because we know the object we are inserting we can know the average
+        // /// case cost if it doesn't already exist
+        // assert_eq!(
+        //     cost.storage_cost.added_bytes,
+        //     average_case_cost.storage_cost.added_bytes
+        // );
+
+        assert_eq!(
+            average_case_cost,
+            OperationCost {
+                seek_count: 8, // todo: why is this 8
+                storage_cost: StorageCost {
+                    added_bytes: 113,
+                    replaced_bytes: 36937, // todo: verify
+                    removed_bytes: NoStorageRemoval,
+                },
+                storage_loaded_bytes: 46420,
+                hash_node_calls: 38, // todo: verify why
+            }
+        );
+    }
+
+    #[test]
+    fn test_batch_average_case_costs() {
+        let db = make_empty_grovedb();
+        let tx = db.start_transaction();
+
+        db.insert(vec![], b"keyb", Element::empty_tree(), None, Some(&tx))
+            .unwrap()
+            .expect("successful root tree leaf insert");
+
+        let ops = vec![GroveDbOp::insert_op(
+            vec![],
+            b"key1".to_vec(),
+            Element::empty_tree(),
+        )];
+        let mut paths = HashMap::new();
+        paths.insert(KeyInfoPath(vec![]), EstimatedLevel(1, AllSubtrees(4, None)));
+        let average_case_cost_result = GroveDb::estimated_case_operations_for_batch(
+            AverageCaseCostsType(paths),
+            ops.clone(),
+            None,
+            |_cost, _old_flags, _new_flags| Ok(false),
+            |_flags, _removed_key_bytes, _removed_value_bytes| {
+                Ok((NoStorageRemoval, NoStorageRemoval))
+            },
+        );
+        assert!(average_case_cost_result.value.is_ok());
+        let cost = db.apply_batch(ops, None, Some(&tx)).cost;
+        // at the moment we just check the added bytes are the same
+        assert_eq!(
+            average_case_cost_result.cost.storage_cost.added_bytes,
+            cost.storage_cost.added_bytes
+        );
     }
 }
