+++ grovedb/src/batch/estimated_costs.rs
@@ -1,4 +1,10 @@
+use std::collections::HashMap;
+
+use merk::estimated_costs::average_case_costs::EstimatedLayerInformation;
+
+use crate::batch::KeyInfoPath;
+
 pub enum EstimatedCostsType {
-    AverageCaseCostsType,
+    AverageCaseCostsType(HashMap<KeyInfoPath, EstimatedLayerInformation>),
     WorstCaseCostsType,
 }
